// apps/dashboard/src/api/generated/restaurantDashboardAPI.schemas.ts
export type OrderSummaryStatus = typeof OrderSummaryStatus[keyof typeof OrderSummaryStatus];


export const OrderSummaryStatus = {
  pending: 'pending',
  preparing: 'preparing',
  ready: 'ready',
  served: 'served',
  completed: 'completed',
  cancelled: 'cancelled',
} as const;

export type OrderSummaryType = typeof OrderSummaryType[keyof typeof OrderSummaryType];


export const OrderSummaryType = {
  'dine-in': 'dine-in',
  takeout: 'takeout',
  delivery: 'delivery',
} as const;

export interface OrderSummary {
  id: string;
  orderNumber: string;
  customerName: string;
  tableNumber: number;
  status: OrderSummaryStatus;
  type: OrderSummaryType;
  total: string;
  itemCount: number;
  createdAt: string;
}

export interface OrderItem {
  id: string;
  orderId: string;
  menuItemId: string | null;
  /** @maxLength 200 */
  name: string;
  /**
     * @minimum -2147483648
     * @maximum 2147483647
     */
  quantity: number;
  unitPrice: string;
  modifiers: string[] | null;
}

export type OrderStatus = typeof OrderStatus[keyof typeof OrderStatus];


export const OrderStatus = {
  pending: 'pending',
  preparing: 'preparing',
  ready: 'ready',
  served: 'served',
  completed: 'completed',
  cancelled: 'cancelled',
} as const;

export type OrderType = typeof OrderType[keyof typeof OrderType];


export const OrderType = {
  'dine-in': 'dine-in',
  takeout: 'takeout',
  delivery: 'delivery',
} as const;

export interface Order {
  id: string;
  /** @maxLength 20 */
  orderNumber: string;
  customerId: string | null;
  /** @maxLength 200 */
  customerName: string;
  /**
     * @minimum -2147483648
     * @maximum 2147483647
     */
  tableNumber: number;
  status: OrderStatus;
  type: OrderType;
  total: string;
  notes: string | null;
  createdAt: string;
  updatedAt: string;
  items: OrderItem[];
}
// ...
export type PatchApiOrdersIdStatusBodyStatus = typeof PatchApiOrdersIdStatusBodyStatus[keyof typeof PatchApiOrdersIdStatusBodyStatus];


export const PatchApiOrdersIdStatusBodyStatus = {
  pending: 'pending',
  preparing: 'preparing',
  ready: 'ready',
  served: 'served',
  completed: 'completed',
  cancelled: 'cancelled',
} as const;